create a separate subscriber to kill server if there is server exception
