+++ src/Server.php
@@ -77,7 +77,6 @@ class Server
             $server->on('message', [$this, 'onMessage']);
         })->otherwise(function ($exception) {
             $this->dispatcher->dispatch(Events::SERVER_EXCEPTION, new ServerExceptionEvent($exception));
-            exit(1);
         });
     }
